+++ docs/CHANGELOG.rst
@@ -13,6 +13,10 @@ Added
 -----
 - Add ``restart`` method to the ``Data`` resource
 
+Changed
+-------
+- Default chunk size for ``iterate`` on ``AnnotationValue`` is set to 10.000
+
 
 ===================
 21.1.0 - 2023-02-09